git-bundle-server: add 'version' command
Add a command to report the version of the compiled bundle server
executables. The version string is set using the '-ldflags' build flag in
'go build' (specifically, setting the 'Version' string in
'cmd/utils/version.go').

In a later patch, the version will be automatically during the build process
(using 'make') with 'git describe'.

Signed-off-by: Victoria Dye <vdye@github.com>