bundle-server: introduce '--auth-config' web server option
Add an '--auth-config' flag to specify the path to a JSON file containing
information used to configure bundle server auth, as well as a function for
parsing the config. For now, no modes are configured, so attempting to pass
a value to the option will fail.

Update the 'git-bundle-web-server' manpage to describe the auth config and
its usage. Also add 'auth-config.md' technical documentation with more
detail around the schema of the auth config file. These will be updated as
new modes are introduced.

Signed-off-by: Victoria Dye <vdye@github.com>