bundle-server: introduce 'authorize' step to 'serve'
Add an 'authorize' function to the bundle web server for managing access to
bundle server content. This function returns one of two states:

- Deny, which indicates that bundle server content should not be served.
  This is configured with the desired 4XX response code and (optional)
  headers. Invoking 'ApplyResult()' with this result will fill in the
  response information and trigger 'serve' to exit immediately.
- Allow, indicating that the requstor may access the requested resource.
  This does not return an immediate response, but if headers are specified,
  they will be added to the eventual response. After applying this result
  with 'ApplyResult()', 'serve' will continue on to add the appropriate
  content to the response.

For now, the 'authorize' function is always nil, so it's never invoked. In
later patches, it will be configured via a new '--auth-config' option to the
web server.

Signed-off-by: Victoria Dye <vdye@github.com>