bundle-server: implement plugin-based auth
Implement and document auth configurations using user-specified runtime
plugins. The goal of this change is to allow users to customize their access
control to a more specific application or domain than is supported by
built-in modes.

A plugin is loaded in 'git-bundle-web-server' from its 'path' by first
comparing the file contents to a specified SHA256 checksum; the web server
fails to start if there is a mismatch. Otherwise, the plugin is loaded and
the specified 'initializer' symbol is looked up. If that symbol exists, the
initializer is called, creating the 'AuthMiddleware' instance and/or an
error. From there, 'git-bundle-web-server' passes the middleware's
'Authorize' function reference to the created 'bundleWebServer' so that it
is invoked after parsing the route of each request.

Additionally, update manpage and technical documentation. Add an example
plugin (built from a standalone '.go' file) & config to 'examples/auth/'.

Signed-off-by: Victoria Dye <vdye@github.com>