test: verify that bundle server started
Update the 'bundleServer.startWebServer()' method to wait for indication
that the web server has successfully started (the "Server is running at
address..." message) before returning successfully. If the process stops
before that, or a timeout of 2s is reached, an error is thrown.

However, because the "Server is running..." printout happens in parallel
with the setup done in 'ListenAndServe[TLS]()', it is possible that the
message is printed and immediately followed with an exit due to a server
startup error. To mitigate that, add a 0.1s pause before printing the
message and, in the tests, only set the status to 'ok' if the 'close' event
has not been triggered by the time the startup message is processed.

Signed-off-by: Victoria Dye <vdye@github.com>